pyproject.toml
[project]
name = "behavioral-sales-coach"
version = "0.1.0"
description = "Voice chatbot with behavioral psychology coaching"
requires-python = ">=3.11"

dependencies = [
    # Core - liquid-audio is the official Liquid AI package
    "liquid-audio",
    "modal>=1.2.4",

    # Audio
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "pyaudio>=0.2.14",
    "pygame>=2.6.0",

    # Utilities
    "numpy>=2.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # Phase II: Semantic matching & UI
    "streamlit>=1.30.0",
    "pinecone-client>=3.0.0",
    "sentence-transformers>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
